Fix pci reader

+++ src/io/io.h
@@ -27,7 +27,7 @@ static inline void out32(uint16_t port, uint32_t data) {
     asm volatile("out %1, %0" : : "a"(data), "d"(port));
 }
 
-static inline uint32_t in32(uint32_t port) {
+static inline uint32_t in32(uint16_t port) {
     uint32_t data;
     asm volatile("in %0, %1": "=a"(data): "d"(port));
     return data;
